Add all hyperscan dependencies as zip files.
These are all the required components.Adding as zip files as a way to avoid http links that might die.  They can also be unzipped where needed by hyperscan.

boost could be used as submodule, but the directory layout changes, and it has to be copied into hyperscan directory anyway.
Both pcre 8.41 and 8.45 are added here in case 8.41 is necessary.  8.45 is last build.
ragel is added as zip here, but could be done as submodule. Using zip to match other dependencies.
sqlite-amalgamation is added as zip, but is only needed for testing.

The only large one here is boost at 90M.